requirements.txt
playwright>=1.41
python-telegram-bot>=21.0
anthropic>=0.23
google-genai>=1.0
duckduckgo-search>=8.0
scipy>=1.12
numpy>=1.26
apscheduler>=3.10
httpx>=0.27
python-dotenv>=1.0
pydantic>=2.6
